+++ src/main/java/com/wedevol/springdatacouchbase/core/service/impl/UserServiceImpl.java
@@ -1,11 +1,20 @@
 package com.wedevol.springdatacouchbase.core.service.impl;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.data.couchbase.config.BeanNames;
+import org.springframework.data.couchbase.core.CouchbaseTemplate;
 import org.springframework.stereotype.Service;
+import com.couchbase.client.java.document.json.JsonObject;
+import com.couchbase.client.java.query.N1qlParams;
+import com.couchbase.client.java.query.N1qlQuery;
+import com.couchbase.client.java.query.N1qlQueryResult;
 import com.wedevol.springdatacouchbase.core.dao.UserCounterRepository;
 import com.wedevol.springdatacouchbase.core.dao.UserRepository;
 import com.wedevol.springdatacouchbase.core.dao.doc.UserBasicDoc;
@@ -31,6 +40,12 @@ public class UserServiceImpl implements UserService {
 
   @Autowired
   private UserCounterRepository userCounterRepo;
+
+  //TODO move this method to a repo class
+  // NOTE: add the qualifier in case you have multiple buckets in your configuration otherwise remove it
+  @Autowired
+  @Qualifier(BeanNames.COUCHBASE_TEMPLATE)
+  private CouchbaseTemplate defaultTemplate; // NOTE: used for the example of N1QL query with cover index
 
   @Override
   public UserDoc findByEmail(String email) {
@@ -111,6 +126,34 @@ public class UserServiceImpl implements UserService {
     List<UserDoc> userDocs = userRepo.deleteUsersWithAge(age);
     logger.info("size: {}", userDocs.size());
     return userDocs;
+  }
+
+  // TODO move this method to a repo class
+  @Override
+  public List<UserBasicDoc> findUsersbyNameUsingTemplateN1QLProjectionWithCoverIndex(String name) {
+    final String cleanName = name.toLowerCase().trim();
+    // NOTE: This method uses Raw N1QL query that projects 1 attribute and it is a covered index ('idx_user_find_by_name')
+    String queryStr = "SELECT u.name FROM users u WHERE u.type = 'com.wedevol.springdatacouchbase.core.dao.doc.UserDoc' AND LOWER(u.name) LIKE '%' || $name || '%'";
+    JsonObject placeholderValues = JsonObject.create().put("name", cleanName);
+    N1qlParams n1qlParams = N1qlParams.build().pretty(false); // TODO add more configurations
+    List<UserBasicDoc> userDocs = defaultTemplate.findByN1QLProjection(N1qlQuery.parameterized(queryStr, placeholderValues, n1qlParams), UserBasicDoc.class);
+    logger.info("size: {}", userDocs.size());
+    return userDocs;
+  }
+
+  //TODO move this method to a repo class
+  @Override
+  public List<Long> findAllUserIdsUsingTemplateN1ql() {
+    // NOTE: This method uses Raw N1QL query that projects 1 attribute
+    String queryStr = "SELECT u.id AS userId FROM users u WHERE u.type = 'com.wedevol.springdatacouchbase.core.dao.doc.UserDoc'";
+    N1qlQueryResult result = defaultTemplate.queryN1QL(N1qlQuery.simple(queryStr));
+    if (!result.errors().isEmpty()) {
+      logger.error("Error running query findAllUserIdsUsingTemplateN1ql()");
+      return Collections.emptyList();
+    }
+    List<Long> userIds = result.allRows().stream().map(row -> row.value().getLong("userId")).collect(Collectors.toList());
+    logger.info("size: {}", userIds.size());
+    return userIds;
   }
 
 }